decouple logic from MarketDetails page container
